requirements.txt
Flask
Flask-SQLAlchemy
Flask-Marshmallow
marshmallow-sqlalchemy
bcrypt
pandas
pytest
selenium
pytest-cov
PyJWT
faker
pyarrow
Flask-WTF
plotly
scikit-learn 
pytest-flask
requests